Move country dropdown generation logic to common utils in core.
