+++ apps/console/src/features/core/utils/index.ts
@@ -23,5 +23,4 @@ export * from "./help-panel-utils";
 export * from "./http-utils";
 export * from "./route-utils";
 export * from "./sort-list";
-export * from "./ui-utils";
 export * from "./user-store-utils";
